# AppWebCanvan/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Kanban API",
        "description": "API para gestionar tableros Kanban con columnas y tareas",
        "contact": {
            "email": "admin@kanban.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/api/columns": {
            "get": {
                "tags": [
                    "Columns"
                ],
                "summary": "Obtener todas las columnas",
                "description": "Retorna una lista de todas las columnas con sus tareas",
                "operationId": "157541de6626c706f141358ff4712aad",
                "responses": {
                    "200": {
                        "description": "Lista de columnas obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "board_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "titulo": {
                                                "type": "string",
                                                "example": "Backlog"
                                            },
                                            "color": {
                                                "type": "string",
                                                "example": "#FF6B6B"
                                            },
                                            "posicion": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "tasks": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "nombre": {
                                                            "type": "string",
                                                            "example": "Configurar autenticación"
                                                        },
                                                        "prioridad": {
                                                            "type": "string",
                                                            "example": "alta"
                                                        },
                                                        "avance": {
                                                            "type": "integer",
                                                            "example": 75
                                                        },
                                                        "status": {
                                                            "type": "string",
                                                            "example": "En curso"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Columns"
                ],
                "summary": "Crear una nueva columna",
                "description": "Crea una nueva columna en un tablero",
                "operationId": "0e8d904832cab548baa6de7884e36cff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "board_id"
                                ],
                                "properties": {
                                    "board_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "titulo": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "En Progreso"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#4ECDC4"
                                    },
                                    "posicion": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Columna creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "board_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "titulo": {
                                            "type": "string",
                                            "example": "En Progreso"
                                        },
                                        "color": {
                                            "type": "string",
                                            "example": "#4ECDC4"
                                        },
                                        "posicion": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/columns/{id}": {
            "get": {
                "tags": [
                    "Columns"
                ],
                "summary": "Obtener una columna específica",
                "description": "Retorna una columna específica con sus tareas",
                "operationId": "7c2769057cc83c33264bd5656723dc4d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la columna",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Columna obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "board_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "titulo": {
                                            "type": "string",
                                            "example": "Backlog"
                                        },
                                        "color": {
                                            "type": "string",
                                            "example": "#FF6B6B"
                                        },
                                        "posicion": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "tasks": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "column_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Configurar autenticación"
                                                    },
                                                    "descripcion": {
                                                        "type": "string",
                                                        "example": "Implementar sistema completo de login"
                                                    },
                                                    "fecha_asignacion": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2025-10-07"
                                                    },
                                                    "fecha_limite": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2025-10-15"
                                                    },
                                                    "asignador": {
                                                        "type": "string",
                                                        "example": "Project Manager"
                                                    },
                                                    "responsable": {
                                                        "type": "string",
                                                        "example": "Juan Pérez"
                                                    },
                                                    "avance": {
                                                        "type": "integer",
                                                        "example": 75
                                                    },
                                                    "prioridad": {
                                                        "type": "string",
                                                        "example": "alta"
                                                    },
                                                    "total_days": {
                                                        "type": "integer",
                                                        "example": 8
                                                    },
                                                    "missing_days": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "En curso"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Columna no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Columns"
                ],
                "summary": "Actualizar una columna",
                "description": "Actualiza la información de una columna específica",
                "operationId": "7ab37e43dfc81a1dfd22bf9e145985ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la columna",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Completado"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#45B7D1"
                                    },
                                    "posicion": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Columna actualizada exitosamente"
                    },
                    "404": {
                        "description": "Columna no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Columns"
                ],
                "summary": "Eliminar una columna",
                "description": "Elimina una columna específica",
                "operationId": "11ae4590fc775f6a137d08724144b17b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la columna",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Columna eliminada exitosamente"
                    },
                    "404": {
                        "description": "Columna no encontrada"
                    }
                }
            }
        },
...
    },
...
}